Threshold selection and raster conversion script
I took a stab at a threshold selection and raster conversion script over the weekend. 

I believe the paper and the code link you uploaded suggested that maximizing selectivity and sensitivity is one of the more accurate selection methods. The function I used was from the pROC library, which allows you to apply that method via the "best" argument and best.method= "youden".

What I didn't do was to cross-validate the selection thresholds to get an average for each species. Thoughts on doing this?? The code wouldn't be difficult to add. But the processing time might take a lot longer.